Helpful RecipeComponent extras (#759)
* Begin adding recipe components with a 'map' operation to operate on a recipe component input/output before it is passed to the parent component

* Register some more components, add helper methods for map and simpleMap,

* Add default Fluid Tag tag component to registry.
Rename constructor to addConstructor for scripters, as constructor is a reserved word in typescript.
Document the map methods in RecipeComponent, as they are intended for scripters not modders.
Improve the dynamic enum component.
Move the default block state components to use vanillas format of a json object, and change the default kubejs one to a special string variant. Note that both will work from scripts no matter what a recipe actually uses, the only thing affected is how KubeJS serializes it. There is no standard agreed upon format for block state objects, so I have used vanillas format and scripters can use simpleMap to change that.